- TorgiService created to handle receiving and processing the GNSS info and allow collection when app is not open - Permissions and initial stability issues addressed - layouts and strings adjusted to support for better cross-device appearance and norms - basic Google Maps integration added
